fix: refactored abbreviation
