+++ services/utils/abbreviateNumbers.js
@@ -1,0 +1,6 @@
+export function abbreviateNumber(n, locale = "en-US") {
+  return Intl.NumberFormat(locale, {
+    notation: "compact",
+    compactDisplay: "short",
+  }).format(n);
+}
