--- util/number.js
@@ -1,6 +1,0 @@
-export function abbreviateNumber(n, locale = "en-US") {
-	return Intl.NumberFormat(locale, {
-		notation: "compact",
-		compactDisplay: "short",
-		}).format(n);
-}
